After much analysis...
It looks like the blue noise in Noise wrapped incorrectly. I was using the byte value + 128, when I should have used the value & 255 . Somehow, the opposite problem seems to apply to the blue noise textures as images (not included here, generated by SquidSquad's tests). It will take some figuring to get to the bottom of this.